# doc/docs/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Theia.cloud API",
    "version" : "0.8.0"
  },
  "paths" : {
    "/service" : {
      "post" : {
        "tags" : [ "Root Resource" ],
        "summary" : "Launch Session",
        "description" : "Launches a session and creates a workspace if required. Responds with the URL of the launched session.",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LaunchRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
    "/service/session" : {
      "post" : {
        "tags" : [ "Session Resource" ],
        "summary" : "Start a new session",
        "description" : "Starts a new session for an existing workspace and responds with the URL of the started session.",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SessionStartRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Session Resource" ],
        "summary" : "Stop session",
        "description" : "Stops a session.",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SessionStopRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "boolean"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      },
      "patch" : {
        "tags" : [ "Session Resource" ],
        "summary" : "Report session activity",
        "description" : "Updates the last activity timestamp for a session to monitor activity.",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SessionActivityRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "boolean"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "deprecated" : true,
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
    "/service/session/performance/{appId}/{sessionName}" : {
      "get" : {
        "tags" : [ "Session Resource" ],
        "summary" : "Get performance metrics",
        "description" : "Returns the current CPU and memory usage of the session's pod.",
        "parameters" : [ {
          "name" : "appId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "sessionName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionPerformance"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
    "/service/session/{appId}/{user}" : {
      "get" : {
        "tags" : [ "Session Resource" ],
        "summary" : "List sessions",
        "description" : "List sessions of a user.",
        "parameters" : [ {
          "name" : "appId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "user",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SessionSpec"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
...
    "/service/{appId}" : {
      "get" : {
        "tags" : [ "Root Resource" ],
        "summary" : "Ping",
        "description" : "Replies if the service is available.",
        "parameters" : [ {
          "name" : "appId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "boolean"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "EnvironmentVars" : {
        "description" : "An object to hold all the ways environment variables can be passed. Not to be used by itself.",
        "type" : "object",
        "properties" : {
          "fromMap" : {
            "description" : "Map of environment variables to be passed to Deployment.  Ignored if Theia applications are started eagerly.  Empty by default.",
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "fromConfigMaps" : {
            "description" : "List of ConfigMaps (by name) containing environment variables to be passed to Deployment as envFrom.configMapRef.  Ignored if Theia applications are started eagerly.  Empty by default.",
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "fromSecrets" : {
            "description" : "List of Secrets (by name) containing environment variables to be passed to Deployment as envFrom.secretRef.  Ignored if Theia applications are started eagerly.  Empty by default.",
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "GitInit" : {
        "description" : "Holds information used to initialize a Workspace with a clone of a Git repository.",
        "required" : [ "repository", "checkout" ],
        "type" : "object",
        "properties" : {
          "repository" : {
            "description" : "The Git repository URL.",
            "type" : "string"
          },
          "checkout" : {
            "description" : "The branch, commit-id, or tag name to checkout.",
            "type" : "string",
            "example" : "main, bd402d6, tags/1.0.0"
          },
          "authInformation" : {
            "description" : "Key for the required auth information, if the repository is not public.",
            "type" : "string"
          }
        }
      },
      "InitOperation" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "arguments" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "LaunchRequest" : {
        "description" : "A request to launch a new session.",
        "required" : [ "appId", "user" ],
        "type" : "object",
        "properties" : {
          "appId" : {
            "description" : "The App Id of this Theia.cloud instance. Request without a matching Id will be denied.",
            "type" : "string"
          },
          "user" : {
            "description" : "The user identification, usually the email address.",
            "type" : "string"
          },
          "appDefinition" : {
            "description" : "The app to launch. Needs to be set if a new or ephemeral session should be launched. For an existing workspace the last app definition will be used if none is given.",
            "type" : "string"
          },
          "workspaceName" : {
            "description" : "The name of the workspace to mount/create. Needs to be set if an existing workspace should be launched.",
            "type" : "string"
          },
          "label" : {
            "description" : "The label of the workspace to mount/create. If no label is given, a default label will be generated.",
            "type" : "string"
          },
          "ephemeral" : {
            "description" : "If true no workspace will be created for the session.",
            "type" : "boolean"
          },
          "timeout" : {
            "format" : "int32",
            "description" : "Number of minutes to wait for session launch. Default is 3 Minutes.",
            "type" : "integer"
          },
          "env" : {
            "allOf" : [ {
              "$ref" : "#/components/schemas/EnvironmentVars"
            }, {
              "description" : "Environment variables"
            } ]
          },
          "gitInit" : {
            "allOf" : [ {
              "$ref" : "#/components/schemas/GitInit"
            }, {
              "description" : "Git Init information"
            } ]
          }
        }
      },
...
      "SessionActivityRequest" : {
        "description" : "A request to report activity for a running session.",
        "required" : [ "appId", "sessionName" ],
        "type" : "object",
        "properties" : {
          "appId" : {
            "description" : "The App Id of this Theia.cloud instance. Request without a matching Id will be denied.",
            "type" : "string"
          },
          "sessionName" : {
            "description" : "The name of the session for which activity is reported.",
            "type" : "string"
          }
        }
      },
...
      "SessionPerformance" : {
        "description" : "Description of the performance of a session",
        "required" : [ "cpuAmount", "cpuFormat", "memoryAmount", "memoryFormat" ],
        "type" : "object",
        "properties" : {
          "cpuAmount" : {
            "description" : "Used CPU amount of the workspace",
            "type" : "string"
          },
          "cpuFormat" : {
            "description" : "Used CPU format of the workspace",
            "type" : "string"
          },
          "memoryAmount" : {
            "description" : "Used memory amount of the workspace",
            "type" : "string"
          },
          "memoryFormat" : {
            "description" : "Used memory format of the workspace",
            "type" : "string"
          }
        }
      },
...
      "SessionSpec" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "appDefinition" : {
            "type" : "string"
          },
          "user" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          },
          "error" : {
            "type" : "string"
          },
          "workspace" : {
            "type" : "string"
          },
          "lastActivity" : {
            "format" : "int64",
            "type" : "integer"
          },
          "sessionSecret" : {
            "type" : "string"
          },
          "envVars" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "envVarsFromConfigMaps" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "envVarsFromSecrets" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "initOperations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/InitOperation"
            }
          }
        }
      },
      "SessionStartRequest" : {
        "description" : "A request to start a session",
        "required" : [ "appId", "user", "appDefinition" ],
        "type" : "object",
        "properties" : {
          "appId" : {
            "description" : "The App Id of this Theia.cloud instance. Request without a matching Id will be denied.",
            "type" : "string"
          },
          "user" : {
            "description" : "The user identification, usually the email address.",
            "type" : "string"
          },
          "appDefinition" : {
            "description" : "The app to launch.",
            "type" : "string"
          },
          "workspaceName" : {
            "description" : "The name of the workspace to mount/create.",
            "type" : "string"
          },
          "timeout" : {
            "format" : "int32",
            "description" : "Number of minutes to wait for session launch. Default is 3 Minutes.",
            "type" : "integer"
          },
          "env" : {
            "allOf" : [ {
              "$ref" : "#/components/schemas/EnvironmentVars"
            }, {
              "description" : "Environment variables"
            } ]
          },
          "gitInit" : {
            "allOf" : [ {
              "$ref" : "#/components/schemas/GitInit"
            }, {
              "description" : "Git Init information"
            } ]
          }
        }
      },
      "SessionStopRequest" : {
        "description" : "A request to stop a session",
        "required" : [ "appId", "user", "sessionName" ],
        "type" : "object",
        "properties" : {
          "appId" : {
            "description" : "The App Id of this Theia.cloud instance. Request without a matching Id will be denied.",
            "type" : "string"
          },
          "user" : {
            "description" : "The user identification, usually the email address.",
            "type" : "string"
          },
          "sessionName" : {
            "description" : "The name of the session to stop.",
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "oauth2",
        "description" : "Authentication",
        "flows" : {
          "implicit" : {
          }
        }
      }
    }
  }
}